moved DormBuilding out of /dorms
